tmdb-cli-tool/requirements.txt
requests
redis
python-dotenv
rich
